Multiple code improvements - squid:S2160, squid:S2184, squid:S2391

+++ src/main/java/org/gedcom4j/relationship/Relationship.java
@@ -87,7 +87,7 @@ public class Relationship implements Comparable<Relationship> {
         if (other == null) {
             return 1;
         }
-        return Math.round(Math.signum(chain.size() - other.chain.size()));
+        return Math.round(Math.signum((float)chain.size() - other.chain.size()));
     }
 
     @Override
